Added a function to match mosaics
imgtools.py is the start of a script for dealing with the images
resulting from emission line fits or projections of the data cube.

There are now 2 functions in imgtools.py.

- gauss2d returns an image with a two dimensional gaussian, with
independent sigmas and a rotation angle.
- match_mosaic currently fits the gauss2d function to an array in
order to match the fluxes of different observations. The original
idea here was to use the broad component as a point source, fit the
gaussian to the broad component flux, and from there be able to
correct the whole cube.