Fixed a bug in linefit
Linefit was getting the wrong spectrum if the data cube was binned and
if individual_spec was set. This bug would also affect the spiral loop,
since the problem was that bin numbers were beeing selected irrespective
of the spaxel coordinates. Now the code searches for the bin number that
matches the coordinates given by the current iteration of xy.