Changes to gmosdc.plotspec
Previous versions of plotspec relied on a try/except methodology to
figure out if the coordinates where of an iterable nature, for when the
user wanted to have an average spectrum between spaxels so and so. This
behaviour has been changed to a more elegant solution based on the
*hasattr* function.

Moreover, this version now allows for combinations of iterable and non
iterable coordinates. For instance, if one wants to have all the
spectra in the second column of spaxels:

mycube.plotspec(1, [0:100])

Additionaly, the averages now work for noise estimates too.