w80eval functional, and velocity based
This commit has a lot of small changes to the w80eval function that
makes it work completely in velocity space. The conversion from
wavelength to velocity is achieved using the astropy.units routines.
There is no transformation for the flux coordinates of the spectrum as
of yet, but perhaps it should be addressed in a future version.