Scale factor and bounds
Previous versions did not allow bounds with *None* values in the
linefit method. This was a limitation caused by an attempt to save a
few lines of code by converting the bounds list to an array and then
multiplying the whole array by the scale factor.

The current fix runs through the bounds list and avoids multiplication
when the bound is a *None*. It works for both limits being *None* or
just one.