Total flux in w80
Previous versions used the maximum value of the cumulative integral as
the total flux in the emission feature. This caused problems when some
of the continuum flux turned out to be negative towards the end of the
spectrum, due to bad continuum subtraction. Now the total flux is
correctly identified as the last value of the cumulative integral
vector.