Major changes in the linefit algorithm.
The main problem with linefit has always been the "Positive derivative
in line search", which basically means that the minimization algorithm
could not find the direction of the gradient. This problem used to
vanish when the variance was removed from the equation.

With the equation written as residuals over variance, any initial guess
that put the model a few tens of times the SNR above the spectrum would
lead to very large values in the chi squared. Naturally, the
minimization algorithm was unable to search for the gradient, when the
scalar to be minimized was some times on the order of millions or more.

This new version takes the square root of the residuals, which greatly
reduces the problem. A future fudge factor for the variance might be
needed, but I am not entirely confortable with that solution.