Linefit, stellar and continuum
Spectrum attributes such as fitspec, resultspec and fitcont are now
predefined as zeros or nans, prior to any validity tests in linefit.

The stellar population spectrum is now treated separately from the
continuum during the whole fitting process. A similar change will be
made to cubetools.gmosdc in a future commit.