Kinematica constraints
Modified parser.py and spectools.py to deal with the kinematic
constraints as set by the k_group parameter in the configuration file.
There is still some work to do, because it seems that the constraints
on the wavelengths are not working properly, although the constraints
on the sigma are correct.