Special sections in parser
Changed the behaviour of the methods parsing the special sections in
the configuration file, such as fit options, continuum options, etc.
Now every check for the presence of the section is made inside the
method itself, instead of in the __load__ method.

Also added a default value for the function value.