Refit with lower flux values
Added a second fit on the linefit that changes the flux values of the
initial guess to the RMS value of the last fit. This feature reduces
the chance that the fit will return high flux values when hitting a
hard iteration limit.

I originally thought of it because many galaxies were getting very high
flux values even when no lines were present.