Outfile name in cubefit
Previous versions let passed the suffix argument to linefit, without
specifying an "outimage". This led to a conflict where the function
that writes the fit results attempted to write the file on the original
directory and not on the current working directory. With the changes in
this commit the cubefit executable performs all the work of setting the
correct output name and sets the "outimage" argument. The original
behaviour of the linefit functions is preserved.