Loading and plotfit
The past few version had an error in the loading routine which caused
the stellar, variance and flag spectrum to not be loaded. The problem
was with trying to set a class attribute via an alias created
dynamically in the loop, like

    for i in (self.a, self.b):
        i = x

Apparently it is not possible to do it like this, so I had to resort to
using

    for i in (self.a, self.b):
        i[:] = x

which seems to be working.

I also fixed the plotfit function, which was not plotting the stellar
component.