Onedspec scale factor
Previous versions of the onedspec linefit function only accepted
positive values for the scale factor. It seemed lika a good idea when I
first wrote it. However, I realized the true purpose of the scale
factor is to bring everything close to unity, therefore it makes no
difference if the spectrum is a little negative, as long as the scale
factor is not negative.

It still requires some more testing for stability, but it makes a lot
more sense now.