Major changes to datacube loading
I made major changes in the loading method of datacube.Cube. The main
reason for this was to have a more standar approach to all the
accessory data (flags, variance, etc.). The previous versions of the
datacube module were heavily based on the original GMOS datacube
structure, which was not very general purpose.