Optimization mask
Previous versions evaluated the line profile function for every pixel
in the spectrum, regardless of it being close to a spectral feature or
not. This resulted in a waste of processing time when large spectra
with few and well spaced lines were fit. To solve this issue I added an
optimization mask to the linefit function, which selects only those
pixels which are within a certain factor of sigmas from the line
center. Sigmas and line centers are taken from the initial guess, and
not updated every time the function has to be evaluated.