Optimization window in linefit
Added an option, accessible via the configuration file, for the
optimization window's width. Also added a few assertion errors to the
fuction that builds the optimization mask.

I took the time to test how much time I gained from performing the
optimization. The results for a spectrum from 3600 to 7000A, with 13
gaussians and a few constraints are:

No optimization = 4min
With optimization = 1min

Quite impressive.