Moved valid_pixels check
The linefit function of onedspec has a test for the share of
valid_pixels in a spectrum. If the more than 80% of the pixels within
th fitting window are flagged or otherwise marked as invalid, then the
function interrupts the fit and exits with status 98. This feature was
already present, this commit simply moved the test to right after the
definition of valid_pixels.