Wavelength in manga
The wavelength vector in the self.wl attribute of manga datacubes
used to be multiplied by 1e+10 to circumvent an inconsistency in
wavelength units. I removed this multiplication for now, becuase
it was causing problems with Califa's data cubes, but some
other solution has to be implemented in the future.