Minimum good pixel fraction
The previous test for the minimum fraction of good pixels was badly
located after the call to the guess parser, which caused errors when
the spectrum was completely flagged. It has now been moved to just
after the check for flags and the definition of the fitting window.
Also, a warning has been added to the code, which informs that the fit
was unsuccessful.

A new argument has been added to Spectrum.linefit, which allows the
user to set the minimum required fraction of good pixels.