Guess parser and valid pixels
The guess parser used to lead to errors when no valid pixels were
available for the basic statistics tests. This caused me to move a test
for the fraction of good pixels to before anything else. However, this
approach led to crashes on datacube when onedspec refused a spectrum.

In this commit I repositioned the good fraction test, and added an
exception catcher on guess_parser, which sets the parameter to np.nan
whenever there is an error trying to get standard deviations, medians,
etc.