Mask file to weights
Once more I find myself troubled by the broad components of the Hydrogen lines. This time, motivated by the difficult
spectrum of Mrk 6, I mustered the will to try a different approach to continuum fitting. Instead of relying on
automatic identification of spectral lines and outliers, I decided it was time to give the user the ability to
arbitrarily define the weights of the pseudo-continuum fit. The solution adopted here is to have a simple function
that reads a mask definition from a file, and applies it to a vector of the same length of the wavelength vector.

Initially all the weights are set to 1. The function then reads each line of the input file, which should have
one interval definition per line, similar to the one below

4800 4900 0
6400 6700 0

Where the first value is the initial wavelength, followed by the final wavelength of the interval in question, and
the weight that should be attributed to all the pixels in this interval.

Since there is no limitation on the value of the weight, this routine can also increase the weight in areas the user
thinks should be favoured by the fitting algorithm.