Added option for SDSS type spectra in specfit.
The files provided by SDSS with a single spectrum come in a particular format, that was not
readily understood by IFSCUBE. This commit adds a subclass of onedspec.Spectrum to deal
with those data, and also adds the option command line option to the specfit executable.

I also changed the nomenclature a little bit, from cubetype to data_type, in the fitter.py
script.