Refactoring of parameter names
Previous versions had LineFit.parameter_names as a list of strings
in the form feature_name.parameter, e.g. 'ha.amplitude'. This was
useful for parsing constraints, but a nuisance when searching for
parameters, as it always involved either a split or format. The
present commit changes this by having the parameter_names as a list
of tuples in the form ('ha', 'amplitude').