Fixed bug in w80
A user reported a bug when evaluating w80 in a datacube. It turns out that cumtrapz
was outputting an array with two dimensions, when one was expected. I do not know
when this started happening, but it seems to be an update to numpy.

The solution was to get only the first index of cumtrapz's output.