Streamlined flux scaling
I moved all the flux scaling to just two methods right before and
after the fit. This removes the many sources of error from carrying
the flux_scale_factor variable around.

I also fixed a bug, and did some refactoring, of the _pack_groups
method, and included the instrument dispersion in the degrees of
freedom estimate.