Renewed velocity width
This commit includes a major change in the way velocity widths
(such as W_80) are calculated. Previous versions used the fit results
directly to get the feature center and integration radius. Although
this was computationally simpler, it required the modeled spectral
feature to be a known function.

In this new version the feature centroid an integration radius is
evaluated based on the cumulative flux of the modeled spectral feature,
allowing for any function, or combination of functions, to serve as
reference. Furthermore, since the model need not be symmetric, the
integration radius now may be different for the upper and lower
limits.

Integration over the observed spectrum is performed after evaluating
the model, and both results are stored in a dictionary, which is then
returned to the LineFit class.

Fixed error in simple fit test.

Also did some minor refactoring of velocity width, and the velocity width plot.

Velocity width working for a single feature.