Added warning for periods in feature naming
The new constraint parser, and lots of other internal steps in modeling,
split the names of features and parameters based on periods (.).
This commit adds a warning if, when loading a previous fit, a feature
name is found to have more than one period.

Also added warning for when extensions are not found in the fit file.