Major rewrite of the channel map function.
Updated the example.

Removed channel_maps import

Since channel maps have been given their on class now, importing the
class here was leading to a circular import statement.
This means that there is no more channel_maps method in the data cube.

Major rewrite of the channel map function.

In this version, the channel_maps function has been made into a class.
The majority of what was the channel_maps function is now the plot
method of the ChannelMaps class.

Significant changes include:
- No longer relying on cubetools.wlprojection for integration.
- Integration over interpolated version of the data, allowing
  the flux of each channel to consider fractions of pixels.

The motivation for this rewrite came from an issue detected by a user,
specifically the lack of flux conservation for different numbers of
channels. In principle, if the velocity limits are kept the same, the
sum over all channel maps should yield the same flux, no matter how
many subdivisions are used. I think I have achieved this behaviour with
the integration over the interpolated data.

Added script for producing fake data.