Fixed guesses for amplitude
There was a terrible bug that affected feature definitions that included peak, mean or median as the amplitude guess.
The problem was that the operations were using the opposite of the mask, i.e. instead of selecting only pixels for which the mask was False, it was selecting only pixels for which the mask was True.

This must have been a leftover from the time before the modeling module.