Updated the fixed fitting method
The fixed fitting method is extremely useful for testing initial guesses for the fit parameters. Up until this point the fixed solution was set at the fitter function, which caused a number of incompatibilities with other methods within the modelling module.
Instead of repeating a lot of code in the fitter.spectrum_fit function I just added another "minimization method" to LineFit.fit called "fixed".
This approach has the advantage of keeping everything exactly equal to what it would be if it were a normal fitting procedure, except for the actual minimization, which instead uses a FixedSolution object with the initial guess as the "x" attribute.