Added options for specifying axes instances for plotting fit results
Previous versions allowed the user to specify the figure, but not the
axes in which to plot the fit results. This has been altered by adding
the optional parameters "spectrum_ax" and "residuals_ax" to the plot
methods in both LineFit and LineFit3D.

This should give the user complete control over these plots.

A side effect of this change is that the "plt.show()" command had
to be moved outside the plot method, which demanded the changes in
the fit_scrutinizer and fitter executables.