Fixed flux conservation in profiles
Previous versions of the 'profiles.f90' functions had
a problem with flux conservation: they were missing
the 1 / (1 + z) factor which should be multiplying
the emitted flux density to yield the observed flux
density.

This version corrects that mistake, and also adds two
tests that perform trapezoidal integration over the
wavelength coordinates to confirm that the integrated
flux of two identical lines at different redshifts
are equal to within a 1e-8 precision.