All velocity_width results in terms of the rest wavelength
Previous versions used to give results for the limiting velocities, as
well as the centroid velocity in terms of the centroid for a given
spectrum.
This has been a source of confusion among users (the author included).
This update changes that philosophy and gives the results all in
reference to the rest wavelength of the spectral feature in question.

As a side effect, it is now mandatory to provide a rest wavelength
in appropriate units.

- Oversampling velocity width

Added an oversample option to velocity_width, in case the user wants
to "increase" the velocity resolution by applying a cubic spline
interpolation to both the model and the data.
This could be used in a future development to return the velocity
corresponding to the peak flux value, without relying too much on
the spectral sampling.

- Updated modelling tests.