Fixed bug caused by not clipping negatives
It turns out that the previous implementation was causing a change of
shape in the data when clipping negatives.
This is caused by a problem when clipping masked_arrays.
The code below illustrates this problem.

'''
x = ma.array(np.arange(5), mask=np.zeros(5).astype(bool))
y = np.clip(x, a_min=0, a_max=10)
'''

After running that, the mask attribute of x will not be directly
passed on to y, but rather y will have its mask converted to a single
False value.
If later the code asks for y[~y.mask], the shape of the resulting
array will be (1, 5), which is the result you get from x[True].
Using ma.clip is no good either, because the current version of numpy
has no dedicated clipping function for masked_arrays.

This commit fixes that mistake by clipping only the data attribute
of the input masked array, and defining a new masked array with the
correct mask from before the clipping.