Added focused plot command line argument
The executables specfit and cubefit always had an option to automatically plot their results, however they lacked an option to focus the plot on the fitting window.
The new '-f', '--focused-plot', command line argument lets the user choose the desired plot format.
No modifications were made to the plotting method itself, the parameter was just "exposed" to the command line.