Streamlined behaviour for clicks on invalid spaxels
The onclick function for selecting spectra in the image plot frame
used to have a lot of try/except clauses, for when the user
selected a spaxel outside the valid area.
It turns out that it is much simpler to add a simple if clause at
the beginning of the function, to check if either the event.xdata and
event.ydata attributes is set to None.