Spaxel exception
An exception in the processing of one spaxel caused a message to be
printed, informing the spaxel, and the associated error message.
This was ok for interactive use, but the correct approach would be
to issue a warning.

A test to check that the correct warning is being issued was added
in this version.

At the same time, two tests that were missing a parameter were fixed.