Differential evolution without constraints
There was a bug with the constraint treatment for fitting with the
differential evolution method.
The loop in the evaluate_constraints function returned an empty
LinearConstraint object when no constraints were provided.
This has been corrected by first checking the length of the list of
constraints, and only entering the LinearConstraint construction
loop when it is greater than zero.

Also added test conditions to check whether the constraints were met
in the fit.