Methods for returning the modelled spectral features
This commit adds a functionality to LineFit and LineFit3D which allows
the user to get a spectrum or data cube of nothing but a few selected
spectral features that were previously fit to the data.
The main motivation is that sometimes it is useful to remove a component
from the fit and attempt to fit the other components without having to
worry about constraints.

This is different from setting a given spectral feature as fixed in the
configuration file, as that spectral feature would have its kinematic
parameters fixed throught the whole data cube.

Development commit messages:

Fixed bug with get_model subclassing

When trying to use the method LineFit3D.velocity_width, there was a
call within the parent method LineFit.velocity_width to self.get_model.
This led to an error, as there is an overriding method
LineFit3D.get_model which was called instead of the 1D method.
The solution was to explicity call the LineFit.get_model method from
within LineFit.velocity_width, instead of using self.get_model.

Extended the get_model function for data cubes.

Added test for more than one spectral feature.

Added function to return the modeled spectrum containing a single
spectral feature.