Fixed bug on accessory data loading
Although FITS extensions can be specified by either an integer or a string, onedspec.Spectrum was ignoring integer specifications without warning.
This fix raises a RuntimeError if the variance, flags or stellar data is specified as something other than a string, an integer or an numpy.ndarray.
It also implements the correct behaviour for when the FITS extension is given as an integer.