Removed ad hoc fix for wavelength units
When astropy's wcs was introduced as the routine for getting the wavelength array from the FITS headers, it led to errors when the resulting array was in meters, instead of angstroms.
A quick and dirty solution for it was introduced in onedspec._load, where the code would check if the keyword 'CUNIT1' was set to meters.
If it were, then the wavelength would be multiplied by 1.0e+10.

This is hardly a solution to the problem, and has been the cause of great confusion among users and developers.
It has now been removed in favour of a better solution which will be implemented soon.