Meters to angstroms
Added a check in the _wavelength methods of Cube and Spectrum to prevent the wavelength from being set in meters. If it detects meters as the WCS unit, it multiplies the resulting wavelength by 1e+10.

This is not the ideal solution yet, partly because the method is repeated in both classes. In a future commit I will try to remove the _wavelength method from the Cube class, and have all the wavelength operations directly inherited from Spectrum.