Channel maps class (#14)
* Removed channel map method

The channel map method in Cube was really just a simple caller function. It is best just to invoke the channel map class directly.

* Moving channel maps to new class

The previous channel maps script mixed plotting and numerical functions in a linear programming style. This new version is an implementation of the ChannelMaps class, in a more object-oriented approach.

* Finishing touches

- Flux units throughout the code.
- Added a write method, which is only a pickle dump.
- Added the channels to the spectrum plot.

* Removed cube from ChannelMaps

The ChannelMaps object was getting much bigger than necessary by carrying the Cube object along with it. This new version removes the ChannelMaps.cube attribute before writing it to disk, making the output file much smaller.

Also made some improvements to the plot spectrum method.

* Minor changes to spectrum plot.

* Changed channel maps example to also load a previous result.

* Code cleanup.

---------

Co-authored-by: danielrd6 <daniel@astro.ufsc.br>