Wavelength fix (#15)
* Removed ad hoc fix for wavelength units

When astropy's wcs was introduced as the routine for getting the wavelength array from the FITS headers, it led to errors when the resulting array was in meters, instead of angstroms.
A quick and dirty solution for it was introduced in onedspec._load, where the code would check if the keyword 'CUNIT1' was set to meters.
If it were, then the wavelength would be multiplied by 1.0e+10.

This is hardly a solution to the problem, and has been the cause of great confusion among users and developers.
It has now been removed in favour of a better solution which will be implemented soon.

* Added assert statements for wavelength ranges

Added two assertion tests to check if the fitting_window limits are compatible with the wavelength coordinates.

* Changed assert statements and added tests

The previous version of the assertion statement required the array to have a strictly ascending orientation.

    fitting_window[0] > wavelength[-1]

instead of the orientation agnostic approach.

    fitting_window[0] > np.max(wavelength)

I guess nobody in their right mind would try to declare the fitting window with the upper limit first.

* Meters to angstroms

Added a check in the _wavelength methods of Cube and Spectrum to prevent the wavelength from being set in meters. If it detects meters as the WCS unit, it multiplies the resulting wavelength by 1e+10.

This is not the ideal solution yet, partly because the method is repeated in both classes. In a future commit I will try to remove the _wavelength method from the Cube class, and have all the wavelength operations directly inherited from Spectrum.

* Removed _wavelength method from Cube

The wavelength method is now solely on the Spectrum class. It seems to be working for both one dimensional spectra and data cubes. I just need a few more tests to make sure everything is where it should be.

* Changed python to 3.9 in environment.yml

This change was motivated by the use of importlib.resources.files in the test_fitter.py script.

* Removed ad hoc fix for wavelength units

When astropy's wcs was introduced as the routine for getting the wavelength array from the FITS headers, it led to errors when the resulting array was in meters, instead of angstroms.
A quick and dirty solution for it was introduced in onedspec._load, where the code would check if the keyword 'CUNIT1' was set to meters.
If it were, then the wavelength would be multiplied by 1.0e+10.

This is hardly a solution to the problem, and has been the cause of great confusion among users and developers.
It has now been removed in favour of a better solution which will be implemented soon.

* Added assert statements for wavelength ranges

Added two assertion tests to check if the fitting_window limits are compatible with the wavelength coordinates.

* Changed assert statements and added tests

The previous version of the assertion statement required the array to have a strictly ascending orientation.

    fitting_window[0] > wavelength[-1]

instead of the orientation agnostic approach.

    fitting_window[0] > np.max(wavelength)

I guess nobody in their right mind would try to declare the fitting window with the upper limit first.

* Removed _wavelength method from Cube

The wavelength method is now solely on the Spectrum class. It seems to be working for both one dimensional spectra and data cubes. I just need a few more tests to make sure everything is where it should be.

* Meters to angstroms

Added a check in the _wavelength methods of Cube and Spectrum to prevent the wavelength from being set in meters. If it detects meters as the WCS unit, it multiplies the resulting wavelength by 1e+10.

This is not the ideal solution yet, partly because the method is repeated in both classes. In a future commit I will try to remove the _wavelength method from the Cube class, and have all the wavelength operations directly inherited from Spectrum.

* Changed python to 3.9 in environment.yml

This change was motivated by the use of importlib.resources.files in the test_fitter.py script.

* Removed double correction for wavelength.

I am still not completely satisfied with this idea of simply multiplying the wavelength by 10^10 when it is in meters, but I think it is at least practical.
The last version had a double correction which has been removed in this commit.

* Removed references to pkg_resources.

The package pkg_resources has been deprecated as an API for getting the files of packages installed in the system, so this commit changed all usages of pkg_resources by importlib.resources.

* Warning filters for tests

Just added a couple of warning filters to the test_wavelength_window functions, because it is expected that many features will be outside the fitting window, and it is not this particular warning that is being tested.

* Finished the wavelength fix.

Added a lengthy docstring to the onedspec.Spectrum._wavelength method, explaining the reasoning behind the transformation from meters to Angstroms. Although the astropy.wcs method has caused some problems, it is much more general than anything I could write. For instance, it can work with the wave-log units of SDSS data seamlessly.

---------

Co-authored-by: danielrd6 <daniel@astro.ufsc.br>