Executable tests are more independent of installation
The previous version of the test_executables.py script required the specfit and cubefit scripts to be present at the PATH, and also be executable by the OS. While this worked fine for typical Linux systems, it failed on Windows because the lack of extension on the script name made the OS unsure on how to run it.

In this version the subprocess.run method invokes the python interpreter explicitly, and passes the absolute path to the script being tested.