Channel maps (#16)
* Changed the spectrum plot

- The spectrum plot is now focused on velocity instead of wavelength.
- Modified the fill_between call to use only the value part of the velocity, continuum and flux quantities. For some reason fill_between does not like instances of units.Quantity, even though plt.plot works fine with it.
- Changed the powers of ten evaluation to work even if there are NaNs and negative values in the spectrum.

* Implemented linear interpolation

The flux integration for the channel maps can be done with linear interpolation and integration via scipy.integrate.quad_vec, instead of only via trapezoidal integration.

* Added basic method for plotting a single map

The plot_map method plots a single channel map. It is useful for a quick look.

* Fixed the number of channel maps

The linspace call in the ChannelMaps.set_channel_boundaries method should have one more step than the desired number of channel maps, since the point here is to get the lower and upper limits of each channel. The central wavelength and velocity coordinates will be computed later.

* Changed operation to get radius in aperture_spectrum.

The previous version implied the same type in the line

    x -= x0

which is complete nonsense as x and y are integers, and x0 and y0 can perfectly well be floats.

* Fixed bug in pseudo continuum fit

The fits where a weights vector were passed directly to the continuum method caused an error because the weights were not subject to the same cut, based on the fitting window, that the data and wavelength were undergoing.

* Changed default options for continuum fitting in channel maps.

* Executable tests are more independent of installation

The previous version of the test_executables.py script required the specfit and cubefit scripts to be present at the PATH, and also be executable by the OS. While this worked fine for typical Linux systems, it failed on Windows because the lack of extension on the script name made the OS unsure on how to run it.

In this version the subprocess.run method invokes the python interpreter explicitly, and passes the absolute path to the script being tested.

* Channel maps deprecation

Added a deprecation warning to the old channelmaps function, and a test to check if the warning is issued.