Removed cube types from fitter
Historically, IFSCube has been developed with the idea that different types of FITS files would be read by different functions. This had the disadvantage of requiring that the user write a particular script to deal with his or hers data.

However, since the implementation of configuration files to set up the fitting process, these scripts are unnecessary. The user simply has to state the extension names (or numbers) where the data reside and IFSCube should be able to work with it. Therefore, the cube_type argument of cubefit or specfit is no longer necessary.

In this commit all references to the cube_type argument have been removed.