Fixed bug in FITS writing
The append_config function of cubetools.py is used to add the configuration file as a table to a new extension in the existing output file from cubefit or specfit. However, instead of opening the FITS file in update (r+) mode and flushing the changes, the file was opened in default (read only) mode, and a call to fits.HDUList.writeto was issued. For some reason this worked in Linux, but luckily Windows had a problem with it.

The new version uses the flush method on a file opened in update mode.