Clear cube types (#17)
* Removed cube types from fitter

Historically, IFSCube has been developed with the idea that different types of FITS files would be read by different functions. This had the disadvantage of requiring that the user write a particular script to deal with his or hers data.

However, since the implementation of configuration files to set up the fitting process, these scripts are unnecessary. The user simply has to state the extension names (or numbers) where the data reside and IFSCube should be able to work with it. Therefore, the cube_type argument of cubefit or specfit is no longer necessary.

In this commit all references to the cube_type argument have been removed.

* Small refactoring.

* Fixed bug in FITS writing

The append_config function of cubetools.py is used to add the configuration file as a table to a new extension in the existing output file from cubefit or specfit. However, instead of opening the FITS file in update (r+) mode and flushing the changes, the file was opened in default (read only) mode, and a call to fits.HDUList.writeto was issued. For some reason this worked in Linux, but luckily Windows had a problem with it.

The new version uses the flush method on a file opened in update mode.

* Removed manga script

The manga.py script was written to work with an in-house version of processed manga files, and was of no general use. Furthermore, it was a source of confusion, causing users to think that this module was required for IFSCube to work with manga data.

* Updated test to point to correct path.

* Removed mentions to GMOS

As was the case with MaNGA, there is no longer any need to provide specific options or modules to deal with GMOS data. Therefore, any references to it have been removed from the code. There are, however, some still lurking around in the documentation and examples.