Fixed environment.yml and a bug in profiles.f90. (#24)
* Changed environment.yml

The previous version had build information associated with each package, which cause problems for installations on different operating systems. This commit fixes this by removing all reference to builds in the package specification.

* Fixed bug in profiles.f90

There was a bug in the declaration of variables in the gauss_hermite_vel subroutine. Line 126 stated

    integer :: n=len(x)

However, this kind of statement is only needed in the comments that f2py reads, in order to specify the correct dimension of the input and output arrays.