Changed scripts to new installation method
Following the move to pyproject.toml, I had to change the old scripts that were kept in the /bin directory to a new form, which can be declared in pyproject.toml, and instruct the installation tool to create executable scripts.

This included defining main functions, and creating the specfit and cubefit functions in ifscube.fitter.