Working version of gauss_vel
At long last, the C version of gauss_vel is working, building correctly with setuptools, and returning the correct results to the python interpreter in the form of a NumPy array. I should work on a appropriate docstring to this function, and then write the remaining three methods for gauss, gauss_hermite, and gauss_hermite_vel. After that I can finally pull this branch into master and make IFSCube compatible with Python 3.12!