Change to NUMBA
The C version of the line profiles did work for some time, but some further testing revealed a flaw when working with multiple components.
After much thought I gave up on trying to write a precompiled function for this task and switched to NUMBA.
NUMBA has the advantage of being completely written in Python and having Just In Time (JIT) compilation, which means that packaging should be easier.
Let's just hope that it remains supported for the next five years or so.