Migrating to numba (#26)
* Changed setup.py from using numpy.distutils to using setuptools.

In this new version the extension module has been modified from a Fortran to a C source.

* Added pure python emission profile module.

During the process of migrating the Fortran module of emission lines profiles to C, I decided to add a pure python module as a stepping stone.
It should also prove useful for users who are unable to compile the extensions for any reason.

* Draft of C module for emission line profiles.

It is not yet completely operational, but it is getting there.

* Fallback import

The modeling module tries to import the compiled version of elprofile, but if it fails, a pure python version is loaded in its place.

* Moved from setup.py to pyproject.toml.

This is the first move towards the new preferred method of packaging.

* Changed scripts to new installation method

Following the move to pyproject.toml, I had to change the old scripts that were kept in the /bin directory to a new form, which can be declared in pyproject.toml, and instruct the installation tool to create executable scripts.

This included defining main functions, and creating the specfit and cubefit functions in ifscube.fitter.

* First working instance of a C extension.

Althought it does not calculate anything useful, it is correctly accepting an array as input.

* Working version of gauss_vel

At long last, the C version of gauss_vel is working, building correctly with setuptools, and returning the correct results to the python interpreter in the form of a NumPy array. I should work on a appropriate docstring to this function, and then write the remaining three methods for gauss, gauss_hermite, and gauss_hermite_vel. After that I can finally pull this branch into master and make IFSCube compatible with Python 3.12!

* C extension for gauss_vel is now fully functional.

Additionaly, I double checked all the math.

* Change to NUMBA

The C version of the line profiles did work for some time, but some further testing revealed a flaw when working with multiple components.
After much thought I gave up on trying to write a precompiled function for this task and switched to NUMBA.
NUMBA has the advantage of being completely written in Python and having Just In Time (JIT) compilation, which means that packaging should be easier.
Let's just hope that it remains supported for the next five years or so.

* Updated requirements.txt

- Removed ppxf and vorbin, and added numba and setuptools.
- Also added updated versions of the packages.

* Changed modeling and tests to work with the new version of the profiles.

The main difference here is that the profile functions used to test whether rest_wl was a scalar or a vector, and then decide what to do with it. Now rest_wl should always be a numpy.ndarray, even if it has only one element.

* Disabled tqdm.

* Re-enabled tqdm.

Disabling tqdm was a bad idea.

* Fixed bug in fit_pseudo_continuum

This commit fixes a clearly nonsensical line which read "if 'weights' is None:", which obviously was meant to be "if weights is None".

* Fixed plotting functions

Since the move to NUMBA all arrays passed to the profile functions should be float64 arrays, and all parameters should be numpy.ndarrays. This commit implements the necessary changes to the plotting method of LineFit.

* Fixed bug in gauss_vel

The profiles.gauss_vel function had a bug.
The equation should use the square of the ratio between observed wavelength and source wavelength, but it was using this ratio to the power of 1.
I also changed a couple of variable names and improved on the readability of the functions.

* Remove setup.py.

* Fixed bug in continuum weights.

The continuum fitting procedure was not masking the weights vector correctly. Both the wavelength and flux vectors were limited to the fitting window, but the weights vector was still the same size as the original spectrum.

* Added dependencies to pyproject.toml.

* Changed trapz to trapezoid and cumtrapz to cumulative_trapezoid.

* Added ppxf to the dependencies.

* Removed manifest and requirement files.

* Removed pPXF wrapper test.

* Fixed bug in continuum weights.

* Changed from cumtrapezoid to cumulative_trapezoid.

* Changed file paths of test files

If you try to run pytest from within the tests directory, which is the default in PyCharm, the tests will now run correctly.

* Remove ppxf requirement and added pytest.

* Fixed arguments passed to gauss_vel.

The numba jit function gauss_vel requires all arguments to be numpy.ndarrays of the float64 type, and the p array, containing the parameters, was all integers.

I took the opportunity to write all function parameters explicitly.

* Remove old_setup.py.